Reuse contract address class
